refactor(server): relocate presentation/watcher modules out of utils
markdown-formatter.ts (pure presentation, 1.2k lines) moves to
src/formatting/; template-renderer.ts (watcher-specific Handlebars) moves to
src/watchers/; trivial vector-math helpers are inlined into their two
consumers and the module deleted. Import paths updated; files moved verbatim.
