+++ packages/server/src/watchers/template-renderer.ts
